The "Ultimate" Fix
The "Genius" I am finally figured out why the merge wouldn't, well, merge. Turns out GitHub detected a space after a comment in one and not in another, which made it think the gates of the nether realm have opened, locking the merge option. The embarrassment is high in this situation. Never make this mistake again future me. ever.